+++ mcpgateway/main.py
@@ -88,6 +88,7 @@ from mcpgateway.observability import init_telemetry
 from mcpgateway.plugins.framework import PluginError, PluginManager, PluginViolationError
 from mcpgateway.routers.server_well_known import router as server_well_known_router
 from mcpgateway.routers.well_known import router as well_known_router
+from mcpgateway.routes.sandbox import router as sandbox_router
 from mcpgateway.schemas import (
     A2AAgentCreate,
     A2AAgentRead,
@@ -7320,6 +7321,7 @@ logger.info(f"Admin API enabled: {ADMIN_API_ENABLED}")
 # Conditional UI and admin API handling
 if ADMIN_API_ENABLED:
     logger.info("Including admin_router - Admin API enabled")
+app.include_router(sandbox_router, prefix="/api/sandbox", tags=["Sandbox"])
     app.include_router(admin_router)  # Admin routes imported from admin.py
 else:
     logger.warning("Admin API routes not mounted - Admin API disabled via MCPGATEWAY_ADMIN_API_ENABLED=False")
